Renamed updateApplication.ts to updater.ts
