+++ src/server/handlers/lifecycle.ts
@@ -1,7 +1,7 @@
 import { HandlerMap, Sender, ok } from "./handlers"
 import { application } from "../../application";
 import { LifeCycleRequest, LifeCycleAction, LifeCyclePingResponse, RequestType } from "../../common/messages";
-import { updateApplication, updateOS } from "../updateApplication";
+import { updateApplication, updateOS } from "../updater";
 
 async function onLifeCycleMessage(request: LifeCycleRequest, send: Sender): Promise<void> {
     switch(request.action) {
